+++ portal/apps/user_requests/views.py
@@ -7,9 +7,8 @@ from django.shortcuts import render, redirect
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework.request import Request
 
-
-
 from portal.apps.error_handling.error_dashboard import new_error
+from portal.apps.error_handling.decorators import handle_error
 from portal.apps.user_requests.api.viewsets import UserRequestViewSet
 from portal.apps.user_requests.models import AerpawUserRequest
 from portal.apps.user_requests.user_requests import approve_user_role_request, deny_user_role_request
@@ -20,6 +19,7 @@ from portal.server.settings import DEBUG, REST_FRAMEWORK
 
 @csrf_exempt
 @login_required
+@handle_error
 def user_role_reqeust_list(request):
     message = None
     try: